Move classifiers to separate files
